+++ tests/test_datetime_helpers.py
@@ -173,7 +173,6 @@ def test_from_rfc3339_w_full_precision():
     assert stamp == expected
 
 
-@staticmethod
 @pytest.mark.parametrize(
     "fractional, nanos",
     [
@@ -281,7 +280,7 @@ def _to_seconds(value):
     """Convert a datetime to seconds since the unix epoch.
 
     Args:
-        value (datetime.datetime): The datetime to covert.
+        value (datetime.datetime): The datetime to convert.
 
     Returns:
         int: Microseconds since the unix epoch.